modularize layers more to hail the arrival of multiple kernels in conv layer
