+++ include/openclAccelerator.h
@@ -24,7 +24,7 @@ public:
     void do_backwardpass_from_output_layer(float* pfExpOut, BaseLossFunction* lossFunc) override;
     void do_backwardpass_dense_layer() override;
 
-    void do_forwardpass_conv_layer(uint input_rows, uint input_columns, uint filter_rows, uint filter_columns) override;
+    void do_forwardpass_conv_layer(sLayer_Dimensions t_tims) override;
 
 
    ~OpenclAccelerator();
