+++ src/Layers/convLayer.cpp
@@ -32,7 +32,7 @@ void ConvLayer::SetPreviousNextLayers(BaseLayer* prevLyr, BaseLayer* nxtLyr)
 void ConvLayer::do_forwardpass_to_current_layer()
 {
     assert(m_bPrevNxtLyrsSet == true);
-    m_pAccelerator->do_forwardpass_conv_layer(m_Dimensions.unInputRows, m_Dimensions.unInputColumns, m_Dimensions.unTransformParametersRows, m_Dimensions.unTransformParametersColumns);
+    m_pAccelerator->do_forwardpass_conv_layer(m_Dimensions);
 }
 
 void ConvLayer::set_transform_matrix_parameter(uint unInIdx, uint unOutIdx, float fWt)
